Discussion Session CSP Solver/requirements.txt
matplotlib
python-constraint
tqdm